+++ Sphinx/ChangeLog.rst
@@ -1,5 +1,8 @@
 ChangeLog
 =========
+* September 6, 2014
+  - Updated media links.
+
 * 01.07.00 --- September 2, 2014
 
   - Added ``plot_scatter()`` function in ``simpleplot`` module.