+++ Sphinx/ChangeLog.rst
@@ -1,6 +1,9 @@
 ChangeLog
 =========
-* September 6, 2014
+* October 4, 2014
+  - Added HSL and HSLA colors format.
+  - Added alternative grey colors.
+
   - Updated media links.
 
 * 01.07.00 --- September 2, 2014