+++ Sphinx/ChangeLog.rst
@@ -1,6 +1,6 @@
 ChangeLog
 =========
-* October 4, 2014
+* 01.08.00 --- October 4, 2014
 
   - Added alternative grey colors.
   - Added HSL and HSLA colors format.